kinda working hit recognition with camera orientation

+++ include/graphics.h
@@ -24,7 +24,7 @@ void		draw(t_window *window);
 void		draw_pixel(t_window *window, int x, int y, uint32_t color);
 void		setup_window(void);
 uint32_t	color(double r, double g, double b, double a);
-double		hit_sphere(t_object sp, t_ray ray);
+int			hit_sphere(t_object sp, t_ray ray);
 
 t_ray		get_ray(t_camera *camera, double x, double y);
